django/requirements.txt
Django
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
Pillow==10.0.1
jsonschema==4.25.1
daphne==4.0.0
channels_redis>=4.0
gunicorn==21.2.0
uvicorn[standard]==0.27.0